requirements.txt
altgraph==0.17.2
astor==0.8.1
Babel==2.9.1
backports.entry-points-selectable==1.1.0
bce-python-sdk==0.8.62
cachetools==4.2.4
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cssselect==1.1.0
cssutils==2.3.0
cycler==0.10.0
decorator==4.4.2
distlib==0.3.3
et-xmlfile==1.1.0
filelock==3.3.0
flake8==4.0.1
Flask==2.0.2
Flask-Babel==2.0.0
future==0.18.2
gast==0.3.3
identify==2.3.0
idna==3.2
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==4.2.0
itsdangerous==2.0.1
Jinja2==3.0.2
kiwisolver==1.3.2
lmdb==1.2.1
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.4.3
mccabe==0.6.1
networkx==2.6.3
nodeenv==1.6.0
numpy==1.19.3
opencv-contrib-python==4.4.0.46
opencv-python==4.5.3.56
openpyxl==3.0.9
paddleocr==2.2.0.2
paddlepaddle==2.1.1
pandas==1.3.3
pefile==2021.9.3
Pillow==8.3.2
platformdirs==2.4.0
pre-commit==2.15.0
premailer==3.10.0
protobuf==3.18.1
pyclipper==1.3.0
pycodestyle==2.8.0
pycryptodome==3.11.0
pyflakes==2.4.0
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
pynput==1.7.3
pyparsing==2.4.7
pypiwin32==223
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytz==2021.3
PyWavelets==1.1.1
pywin32==302
pywin32-ctypes==0.2.0
PyYAML==5.4.1
requests==2.26.0
scikit-image==0.17.2
scipy==1.7.1
Shapely==1.7.1
shellcheck-py==0.7.2.1
six==1.16.0
tifffile==2021.10.10
toml==0.10.2
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
virtualenv==20.8.1
visualdl==2.2.1
Werkzeug==2.0.2
zipp==3.6.0
